+++ swagger_parser/test/e2e/tests/basic/basic_requests.3.0/expected_files/clients/auth_client.dart
@@ -1,0 +1,20 @@
+// coverage:ignore-file
+// GENERATED CODE - DO NOT MODIFY BY HAND
+// ignore_for_file: type=lint, unused_import
+
+import 'package:dio/dio.dart';
+import 'package:retrofit/retrofit.dart';
+
+import '../models/register_user_dto.dart';
+
+part 'auth_client.g.dart';
+
+@RestApi()
+abstract class AuthClient {
+  factory AuthClient(Dio dio, {String? baseUrl}) = _AuthClient;
+
+  @POST('/api/Auth/register')
+  Future<void> postApiAuthRegister({
+    @Body() RegisterUserDto? body,
+  });
+}