+++ swagger_parser/test/e2e/tests/basic/reference_types_class.3.0/expected_files/export.dart
@@ -1,0 +1,7 @@
+// coverage:ignore-file
+// GENERATED CODE - DO NOT MODIFY BY HAND
+// ignore_for_file: type=lint, unused_import
+
+// Data classes
+export 'models/class_name.dart';
+export 'models/another_class.dart';