+++ swagger_parser/test/e2e/tests/corrector/expected_files/clients/client_client.dart
@@ -6,15 +6,19 @@ import 'dart:convert';
 import 'package:dio/dio.dart';
 import 'package:retrofit/retrofit.dart';
 
-import '../models/class_name1.dart';
-import '../models/class_name2.dart';
-import '../models/class_name3.dart';
-import '../models/class_name5.dart';
-import '../models/data1_att.dart';
-import '../models/data_att2.dart';
-import '../models/data_att4.dart';
-import '../models/private_att_data3.dart';
-import '../models/private_class_name4.dart';
+import '../models/class_name1_dto.dart';
+import '../models/class_name2_dto.dart';
+import '../models/class_name3_dto.dart';
+import '../models/class_name56_dto.dart';
+import '../models/class_name5_dto.dart';
+import '../models/data1_att_dto.dart';
+import '../models/data_att2_dto.dart';
+import '../models/data_att4_dto.dart';
+import '../models/enum0_dto.dart';
+import '../models/object0_dto.dart';
+import '../models/object1_dto.dart';
+import '../models/private_att_data3_dto.dart';
+import '../models/private_class_name4_dto.dart';
 
 part 'client_client.g.dart';
 
@@ -22,17 +26,29 @@ part 'client_client.g.dart';
 abstract class ClientClient {
   factory ClientClient(Dio dio, {String? baseUrl}) = _ClientClient;
 
+  /// [p1Class] - Name not received and was auto-generated.
+  ///
+  /// [p2Enum] - Name not received and was auto-generated.
   @MultiPart()
   @GET('/api/v1/')
-  Future<String> apiV1({
-    @Part(name: 'p1') required ClassName1 p1,
-    @Part(name: 'p2') required ClassName2 p2,
-    @Part(name: 'p3') required ClassName3 p3,
-    @Part(name: 'p4') required PrivateClassName4 p4,
-    @Part(name: 'p5') required ClassName5 p5,
-    @Part(name: 'v1') required Data1Att v1,
-    @Part(name: 'v2') required DataAtt2 v2,
-    @Part(name: 'v3') required PrivateAttData3 v3,
-    @Part(name: 'v4') required DataAtt4 v4,
+  Future<String> getApiV1({
+    @Part(name: 'p1') required ClassName1Dto p1,
+    @Part(name: 'p3') required ClassName3Dto p3,
+    @Part(name: 'p5') required ClassName5Dto p5,
+    @Part(name: 'v1') required Data1AttDto v1,
+    @Part(name: 'v3') required PrivateAttData3Dto v3,
+    @Part(name: 'p2') ClassName2Dto? p2,
+    @Part(name: 'p4') PrivateClassName4Dto? p4,
+    @Part(name: 'p6') ClassName56Dto? p6,
+    @Part(name: 'v2') DataAtt2Dto? v2,
+    @Part(name: 'v4') DataAtt4Dto? v4,
+    @Part(name: 'p1_class') Object0Dto? p1Class,
+    @Part(name: 'p2_enum') Enum0Dto? p2Enum,
+  });
+
+  /// [body] - Name not received and was auto-generated.
+  @POST('/api/v1/')
+  Future<String> postApiV1({
+    @Body() Object1Dto? body,
   });
 }