+++ swagger_parser/test/e2e/tests/corrector/expected_files/models/class_name5.dart
@@ -1,5 +1,0 @@
-// coverage:ignore-file
-// GENERATED CODE - DO NOT MODIFY BY HAND
-// ignore_for_file: type=lint, unused_import
-
-typedef ClassName5 = dynamic;